remove middleware files based on HTTP method, we don't have enough functions to make the files ridiculously large. Consolidate all functions into userMiddleware.js
