+++ middleware/sharedLinkMiddleware.js
